+++ platform/swift/source/LoggerBridge.swift
@@ -85,6 +85,7 @@ final class LoggerBridge: LoggerBridging {
         network: Network?,
         errorReporting: RemoteErrorReporting,
         sleepMode: SleepMode,
+        initialFields: [CapturePassable.Field],
         issueCallbackConfiguration: IssueCallbackConfiguration?
     ) {
         do {
@@ -111,6 +112,7 @@ final class LoggerBridge: LoggerBridging {
             network,
             errorReporting,
             sleepMode == SleepMode.enabled,
+            initialFields,
             issueCallbackConfiguration
         )
 
@@ -145,6 +147,7 @@ final class LoggerBridge: LoggerBridging {
         network: Network?,
         errorReporting: RemoteErrorReporting,
         sleepMode: SleepMode,
+        initialFields: [CapturePassable.Field],
         issueCallbackConfiguration: IssueCallbackConfiguration?
     ) -> LoggerBridging? {
         return LoggerBridge(
@@ -163,6 +166,7 @@ final class LoggerBridge: LoggerBridging {
             network: network,
             errorReporting: errorReporting,
             sleepMode: sleepMode,
+            initialFields: initialFields,
             issueCallbackConfiguration: issueCallbackConfiguration
         )
     }