+++ test/benchmark/src/bin/logger_benchmark.rs
@@ -58,6 +58,7 @@ fn simple_log(c: &mut Criterion) {
       ..Default::default()
     }),
     initial_ootb_fields: [].into(),
+    initial_custom_fields: [].into(),
     resource_utilization_target: Box::new(bd_test_helpers::resource_utilization::EmptyTarget),
     session_replay_target: Box::new(bd_test_helpers::session_replay::NoOpTarget),
     events_listener_target: Box::new(bd_test_helpers::events::NoOpListenerTarget),
@@ -101,6 +102,7 @@ fn with_matcher_and_buffer(c: &mut Criterion) {
       ..Default::default()
     }),
     initial_ootb_fields: [].into(),
+    initial_custom_fields: [].into(),
     resource_utilization_target: Box::new(bd_test_helpers::resource_utilization::EmptyTarget),
     session_replay_target: Box::new(bd_test_helpers::session_replay::NoOpTarget),
     events_listener_target: Box::new(bd_test_helpers::events::NoOpListenerTarget),