+++ test/platform/swift/unit_integration/core/Logger+Tests.swift
@@ -15,7 +15,8 @@ extension Logger {
         withAPIKey apiKey: String = "",
         sessionStrategy: SessionStrategy = .fixed(),
         dateProvider: DateProvider? = nil,
-        fieldProviders: [FieldProvider] = [],
+        customFieldGetters: [MetadataProviderController.FieldGetter] = [],
+        initialFields: Fields = [:],
         configuration: Configuration = .testConfiguration,
         loggerBridgingFactoryProvider: LoggerBridgingFactoryProvider = LoggerBridgingFactory()
     ) throws -> Logger
@@ -27,7 +28,8 @@ extension Logger {
                 configuration: configuration,
                 sessionStrategy: sessionStrategy,
                 dateProvider: dateProvider,
-                fieldProviders: fieldProviders,
+                customFieldGetters: customFieldGetters,
+                initialFields: initialFields,
                 storageProvider: MockStorageProvider(),
                 timeProvider: SystemTimeProvider(),
                 loggerBridgingFactoryProvider: loggerBridgingFactoryProvider